+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.2.2] - 2024/04/29
+
+### Updated
+
+- Building instructions (adding a section for cutlass)
+- Checksums for custom torch builds (within docker)
+
+### Fixed
+
+- An error in `pack_fp_weight`
+
 ## [0.2.1] - 2024/04/27
 
 ### Fixed